+++ streampipes-processors-transformation-jvm/src/main/java/org/apache/streampipes/processors/transformation/jvm/processor/state/labeler/buffer/StateBufferLabelerController.java
@@ -97,7 +97,7 @@ public class StateBufferLabelerController extends StandaloneEventProcessingDecla
 
     String labelName = getLabelName(extractor);
 
-    List<Integer> numberValues = getNumberValues(extractor);
+    List<Double> numberValues = getNumberValues(extractor);
 
     List<String> labelStrings = getLabelStrings(extractor);
